requirements.txt
absl-py==1.4.0
autoattack==0.1
boltons==23.0.0
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chardet==4.0.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cmake==3.26.0
conda==23.1.0
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1673298753778/work
cycler==0.11.0
filelock==3.10.0
fonttools==4.39.2
geotorch==0.3.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
huggingface-hub==0.13.2
idna==2.10
Jinja2==3.1.2
kiwisolver==1.4.4
lit==15.0.7
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.0
ninja==1.11.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
packaging==23.0
pandas==1.3.5
patsy==0.5.3
Pillow==9.4.0
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.25.1
requests-oauthlib==1.3.1
robustbench==1.1
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
scipy==1.10.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
statsmodels==0.13.5
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
timm==0.6.12
toolz @ file:///croot/toolz_1667464077321/work
torch==2.0.0
torchaudio==2.0.1
torchdiffeq==0.2.3
torchsde==0.2.5
torchvision==0.15.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
trampoline==0.1.2
triton==2.0.0
typing_extensions==4.5.0
urllib3 @ file:///croot/urllib3_1673575502006/work
Werkzeug==2.2.3
zstandard @ file:///opt/conda/conda-bld/zstandard_1663827383994/work
